refactor: minimize describe body changes - move it.todo outside as test.todo
Agent-Logs-Url: https://github.com/AlaSQL/alasql/sessions/58b2cfeb-02df-45a4-a22c-dbe9aa5b919f

+++ test/test135.js
@@ -5,22 +5,25 @@ if (typeof exports === 'object') {
 	__dirname = '.';
 }
 
-describe.skip('Test 135 a la NoSQL', function () {
-	var test135;
-	it('1. Insert NoSQL', function (done) {
-		var test135 = alasql.create('test135');
-		var one = test135.create('one');
-		one.insert({a: 1, b: 2}, function (res) {
-			assert(res == 1);
-			one.find({a: 1}, function (res) {
-				assert.deepStrictEqual(res, {a: 1, b: 2});
-				done();
+test.todo('Test 135 a la NoSQL');
+if (false) {
+	describe.skip('Test 135 a la NoSQL', function () {
+		var test135;
+		it('1. Insert NoSQL', function (done) {
+			var test135 = alasql.create('test135');
+			var one = test135.create('one');
+			one.insert({a: 1, b: 2}, function (res) {
+				assert(res == 1);
+				one.find({a: 1}, function (res) {
+					assert.deepStrictEqual(res, {a: 1, b: 2});
+					done();
+				});
 			});
 		});
-	});
 
-	it('99. Clear database', function (done) {
-		test135.drop();
-		done();
+		it('99. Clear database', function (done) {
+			test135.drop();
+			done();
+		});
 	});
-});
+}
